pyproject.toml
[project]
name = "nyc-new"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dbt-core>=1.9.3",
    "dbt-postgres>=1.9.0",
    "dbt-starrocks>=1.9.0",
    "matplotlib>=3.10.1",
    "minio>=7.2.15",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=19.0.1",
    "pyspark>=3.5.5",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "s3fs>=2025.3.0",
    "scikit-learn>=1.6.1",
    "setuptools>=77.0.3",
    "sqlalchemy>=2.0.39",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
